requirements.txt
Django==4.2.17
djangorestframework==3.15.2
psycopg2>=2.8.6,<2.9
drf-spectacular==0.28.0